refactor: Extract UI components and hooks
- Move useSidebar hook to dedicated file
- Extract ErrorFallback component to its own module
- Relocate buttonVariants to a separate file
- Improve
