+++ frontend/src/hooks/ui/useSidebar.ts
@@ -1,0 +1,8 @@
+import { useContext } from 'react';
+import { SidebarContext } from '@/context/SidebarContext';
+
+function useSidebar() {
+    return useContext(SidebarContext);
+}
+
+export { useSidebar };
