+++ iviss-backend/src/main.rs
@@ -1,3 +1,4 @@
+use anyhow::Context;
 use iviss_backend::api_doc::ApiDoc;
 use iviss_backend::app_state::AppState;
 use iviss_backend::config::{Config, Environment};
@@ -5,7 +6,6 @@ use iviss_backend::db::initialize_pool;
 use iviss_backend::db::initialize_redis_pool;
 use iviss_backend::routes;
 use iviss_backend::services::sms_provider::{MockSmsProvider, SmsProvider, TwilioSmsProvider};
-use anyhow::Context;
 use std::net::SocketAddr;
 use std::sync::Arc;
 use tracing::info;
