+++ iviss-backend/src/tests/mod.rs
@@ -1,7 +1,5 @@
-#[cfg(test)]
-pub mod activation_service_tests;
-#[cfg(test)]
-pub mod terminate_session_tests;
 pub mod jwt_service_tests;
 #[cfg(test)]
 pub mod otp_service_tests;
+#[cfg(test)]
+pub mod terminate_session_tests;
